refactor PSPPanel and Defect Log code
