+++ src/net/sf/memoranda/ui/PSPPanel.java
@@ -4,41 +4,21 @@ import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Font;
-import java.awt.Point;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-import java.util.Vector;
 
 import javax.swing.DefaultCellEditor;
 import javax.swing.JButton;
-import javax.swing.JEditorPane;
-import javax.swing.JFileChooser;
-import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JComboBox;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
 import javax.swing.JToolBar;
-import javax.swing.table.AbstractTableModel;
 import javax.swing.table.TableColumn;
-import javax.swing.SwingUtilities;
-import javax.swing.event.HyperlinkEvent;
-import javax.swing.event.HyperlinkListener;
-import javax.swing.JOptionPane;
+import javax.swing.table.TableModel;
 
-import javax.swing.table.AbstractTableModel;
-import javax.swing.table.TableColumn;
-import org.knowm.xchart.CategoryChart;
-import org.knowm.xchart.CategoryChartBuilder;
-import org.knowm.xchart.Histogram;
-import org.knowm.xchart.XChartPanel;
-import org.knowm.xchart.style.Styler.LegendPosition;
 import net.sf.memoranda.CurrentProject;
-import net.sf.memoranda.Defect;
 import net.sf.memoranda.DefectList;
 import net.sf.memoranda.History;
 import net.sf.memoranda.NoteList;
@@ -47,19 +27,18 @@ import net.sf.memoranda.Project;
 import net.sf.memoranda.ProjectListener;
 import net.sf.memoranda.ResourcesList;
 import net.sf.memoranda.TaskList;
-import net.sf.memoranda.date.CalendarDate;
+import net.sf.memoranda.ui.DefectPanel.DefectsTableModel;
 import net.sf.memoranda.util.Local;
 
 public class PSPPanel extends JPanel {
-	/**
-	 * 
-	 */
+	
 	private static final long serialVersionUID = 1L;
 	BorderLayout borderLayout = new BorderLayout();
   
 	JToolBar toolBar = new JToolBar();
 	JButton historyBackB = new JButton();
 	JButton historyForwardB = new JButton();
+	//Defect Log Buttons
 	JButton addRowBtn = new JButton("Add Row");
 	JButton deleteBtn = new JButton("Delete");
 
@@ -75,178 +54,6 @@ public class PSPPanel extends JPanel {
     JButton start = new JButton ("Start");
     JButton reset = new JButton ("Reset");
 	
-	/**
-	 * Defect Log Table Model
-	 */
-	public class DefectsTableModel extends AbstractTableModel {
-		/**
-		 * 
-		 */
-		private static final long serialVersionUID = 1L;
-
-		// Table data
-		private ArrayList<Object[]> _data;
-
-		// Next defect number
-		private int _nextNumber;
-
-		// Column names
-		private String[] ColumnNames = {
-				"Number",
-				"Date Found",
-				"Type",
-				"Phase Injected",
-				"Phase Corrected",
-				"Fix Time (mins)",
-				"Description"
-		};
-
-		// Column types
-		private Class[] ColumnClasses = {
-				Integer.class,
-				CalendarDate.class,
-				String.class,
-				String.class,
-				String.class,
-				Integer.class,
-				String.class
-		};
-
-		public DefectsTableModel() {
-			_data = new ArrayList<Object[]>();
-			_nextNumber = 1;
-		}
-
-		@Override
-		public int getRowCount() {
-			return _data.size();
-		}
-
-		@Override
-		public int getColumnCount() {
-			return ColumnNames.length;
-		}
-
-		@Override
-		public Object getValueAt(int row, int column) {
-			return _data.get(row)[column];
-		}
-
-		@Override
-		public Class getColumnClass(int c) {
-			return ColumnClasses[c];
-		}
-
-		@Override
-		public String getColumnName(int c) {
-			return ColumnNames[c];
-		}
-
-		@Override
-		public boolean isCellEditable(int row, int col) {
-			return (col != 0 && col != 1);
-		}
-
-		@Override
-		public void setValueAt(Object val, int row, int col) {
-			_data.get(row)[col] = val;
-
-			String attrib = "";
-			String value = "";
-			if (col == 2)
-			{
-				attrib = "type";
-			}
-			else if (col == 3)
-			{
-				attrib = "injected";
-			}
-			else if (col == 4)
-			{
-				attrib = "corrected";
-			}
-			else if (col == 5)
-			{
-				attrib = "fix_time";
-				value = Integer.toString((Integer)val);
-			}
-			else if (col == 6)
-			{
-				attrib = "description";
-			}
-
-			if (col != 5)
-			{
-				value = (String)val;
-			}
-
-			CurrentProject.getDefectList().modifyDefect((Integer)_data.get(row)[0], attrib, value);
-			fireTableDataChanged();
-		}
-
-		/**
-		 * Add a new row for a new defect: has default data inputs.
-		 */
-		public void addNewRow() {
-			_data.add(new Object[] {
-					_nextNumber,
-					new CalendarDate(),
-					"",
-					"",
-					"",
-					0,
-					""
-			});
-
-			Defect temp_defect = new Defect();
-			temp_defect.SetDateFound(CalendarDate.today());
-			temp_defect.SetDefectNumber(_nextNumber++);
-			CurrentProject.getDefectList().addDefect(temp_defect);
-
-			fireTableDataChanged();
-		}
-
-		/**
-		 * Deletes the specified rows from the table
-		 */
-		public void deleteRows(int[] rowIndexes) {
-			System.out.println(rowIndexes.length);
-			for(int i = rowIndexes.length - 1; i >= 0; i--) {
-				CurrentProject.getDefectList().removeDefect((Integer)_data.get(rowIndexes[i])[0]);
-				_data.remove(rowIndexes[i]);
-			}
-
-			fireTableDataChanged();
-		}
-
-		/**
-		 * Creates table with existing defects
-		 */
-		protected void populateTable()
-		{
-			_data.clear();
-			Vector<Defect> defects = CurrentProject.getDefectList().getAllDefect();
-			for (int i = 0; i < defects.size(); ++i)
-			{
-
-				_data.add(new Object[] {
-						defects.get(i).GetDefectNumber(),
-						defects.get(i).GetDateFound(),
-						defects.get(i).GetType(),
-						defects.get(i).GetInjected(),
-						defects.get(i).GetCorrected(),
-						defects.get(i).GetFixTime(),
-						defects.get(i).GetDescription()
-				});
-				if (i == defects.size() - 1)
-				{
-					_nextNumber = defects.get(i).GetDefectNumber() + 1;
-				}
-			}
-			fireTableDataChanged();
-		}
-	}
-
 	// Things inside the panel that people need to see!
 	public PSPPanel(DailyItemsPanel _parentPanel){
 		try {
@@ -297,13 +104,15 @@ public class PSPPanel extends JPanel {
 
 		toolBar.add(addRowBtn);
 		toolBar.add(deleteBtn);
-		this.add(toolBar, BorderLayout.NORTH);		
-	
-		DefectsTableModel tableModel = new DefectsTableModel();
+		this.add(toolBar, BorderLayout.NORTH);	
+		
+		//Defect Log Table Creator	
+		DefectPanel defectLog = new DefectPanel();
+		TableModel tableModel = defectLog.new DefectsTableModel();
 		JTable table = new JTable(tableModel);
 		JScrollPane scrollPane2 = new JScrollPane(table);
 		
-		// Table Size
+		// Defect Log Table Size
 	
 		table.getColumnModel().getColumn(0).setPreferredWidth(20*1000);		
 		table.getColumnModel().getColumn(1).setPreferredWidth(50*1000);
@@ -355,19 +164,19 @@ public class PSPPanel extends JPanel {
 		this.add(scrollPane2, BorderLayout.CENTER);
 
 		table.setFillsViewportHeight(true);
-		tableModel.populateTable();
+		((DefectsTableModel) tableModel).populateTable();
 
 		// The Add Row button adds a new row to the table
 		addRowBtn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
-				tableModel.addNewRow();
+				((DefectsTableModel) tableModel).addNewRow();
 			}
 		});
 
 		// The Delete button deletes the selected rows
 		deleteBtn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
-				tableModel.deleteRows(table.getSelectedRows());
+				((DefectsTableModel) tableModel).deleteRows(table.getSelectedRows());
 			}
 		});
 		
@@ -376,7 +185,7 @@ public class PSPPanel extends JPanel {
 
 			}
 			public void projectWasChanged() {
-				tableModel.populateTable();
+				((DefectsTableModel) tableModel).populateTable();
 			}
 		});
